+++ src/Common/Types.fs
@@ -4,6 +4,9 @@ open Kafka
 open Metrics
 open ServiceIdentification
 
+[<Measure>] type second
+[<Measure>] type attempt
+
 //
 // Metrics
 //
@@ -12,6 +15,12 @@ type CustomMetric = {
     Name: MetricName
     Type: MetricType
     Description: string
+}
+
+type ResourceMetricInInterval = {
+    Resource: ResourceAvailability
+    Interval: int<second>
+    Checker: unit -> ResourceStatus
 }
 
 type MetricsRoute = private MetricsRoute of string
@@ -84,9 +93,6 @@ module Connections =
 //
 
 // Handlers and policies
-
-[<Measure>] type second
-[<Measure>] type attempt
 
 type ErrorMessage = string
 
@@ -277,6 +283,7 @@ type ConfigurationParts<'InputEvent, 'OutputEvent> = {
     FromDomain: Map<ConnectionName, FromDomain<'OutputEvent>>
     MetricsRoute: MetricsRoute option
     CustomMetrics: CustomMetric list
+    IntervalResourceCheckers: ResourceMetricInInterval list
     CreateInputEventKeys: CreateInputEventKeys<'InputEvent> option
     CreateOutputEventKeys: CreateOutputEventKeys<'OutputEvent> option
     KafkaChecker: Checker option
@@ -302,6 +309,7 @@ module internal ConfigurationParts =
             FromDomain = Map.empty
             MetricsRoute = None
             CustomMetrics = []
+            IntervalResourceCheckers = []
             CreateInputEventKeys = None
             CreateOutputEventKeys = None
             KafkaChecker = None
@@ -331,6 +339,7 @@ type KafkaApplicationParts<'InputEvent, 'OutputEvent> = {
     ProducerErrorHandler: ProducerErrorHandler
     MetricsRoute: MetricsRoute option
     CustomMetrics: CustomMetric list
+    IntervalResourceCheckers: ResourceMetricInInterval list
     PreparedRuntimeParts: PreparedConsumeRuntimeParts<'OutputEvent>
 }